Get predictions from a GAM model.

It returns a tibble with the predictions from a a gam or bam object.

get_gam_predictions(
  model,
  series,
  series_length = 25,
  conditions = NULL,
  exclude_random = TRUE,
  exclude_terms = NULL,
  split = NULL,
  sep = "\\.",
  time_series,
  transform = NULL,
  ci_z = 1.96,
  .comparison = NULL
)

Arguments

model
A gam or bam model object.

series
An unquoted expression indicating the model term that defines the series on which smoothing is applied. This is the term that is displayed on the x-axis when plotting.

series_length
An integer indicating how many values along the time series to use for predicting the outcome term.

conditions
A list of quosures with quos specifying the levels to plot from the model terms.

exclude_random
Whether to exclude random smooths (the default is TRUE).

exclude_terms
Terms to be excluded from the prediction. Term names should be given as they appear in the model summary (for example, "s(x0,x1)").

split
Columns to separate as a named list.

sep
Separator between columns (default is "\.", which is the default with ). If character, it is interpreted as a regular expression.

time_series
Deprecated, use series instead.

transform
Function used to transform the fitted values (useful for getting plots on the response scale).

ci_z
The z-value for calculating the CIs (the default is 1.96 for 95 percent CI).

.comparison
Internal parameter, passed from plot_smooths().

Examples

library(mgcv)
set.seed(10)
data <- gamSim(4)

#> Factor `by' variable example

model <- gam(y ~ fac + s(x2) + s(x2, by = fac) + s(x0), data = data)

pred <- get_gam_predictions(model, x2)